requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
pydantic>=2.5.0
pysyncobj>=0.3.13
